+++ packages/components/atoms/f-button/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
 const Button = require('../../test-utils/component-objects/f-button.component');
@@ -11,10 +10,8 @@ describe('Accessibility tests', () => {
         button = new Button();
         button.withQuery('knob-Button Type', 'primary');
         button.withQuery('knob-Button Size', 'medium');
-        const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-        button.open(pageUrl);
-
-        button.waitForActionComponent();
+        button.load();
+        
         const axeResults = getAccessibilityTestResults('f-button - action');
 
         expect(axeResults.violations.length).toBe(0);
@@ -27,10 +24,8 @@ describe('Accessibility tests', () => {
             .withQuery('knob-href', 'link')
             .withQuery('knob-Button Size', 'medium');
 
-        const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-        button.open(pageUrl);
-
-        button.waitForLinkComponent();
+        button.load("link");
+        
         const axeResults = getAccessibilityTestResults('f-button - link');
 
         // Assert