+++ packages/components/molecules/f-tabs/test-utils/component-objects/f-tabs.component.js
@@ -9,6 +9,10 @@ module.exports = class Tabs extends Page {
 
     get tabButtons () { return $$('[data-test-id*="tab-button"]'); }
 
+    load () {
+        super.load(this.component);
+    }
+
     open (url) {
         super.open(url);
     }