+++ packages/components/organisms/f-checkout/src/tenants/en-AU.js
@@ -10,7 +10,7 @@ const messages = {
         administrativeArea: 'State',
         postcode: 'Postcode',
         deliveryOrderMethod: 'Delivery time',
-        collectionOrderMethod: 'Collection time',
+        collectionOrderMethod: 'Pick-up time',
         dineinOrderMethod: 'Dine in time',
         tableIdentifier: 'Table number or name'
     },
@@ -112,7 +112,7 @@ const messages = {
 
             LOCATION_UNDELIVERABLE: {
                 title: 'Cannot deliver to address',
-                message: 'We’re sorry, but the restaurant does not currently deliver to this address. Please choose to order via collection or from another restaurant.',
+                message: 'We’re sorry, but the restaurant does not currently deliver to this address. Please choose to order via pick-up or from another restaurant.',
                 buttonText: 'Ok'
             },
 
@@ -126,6 +126,12 @@ const messages = {
                 title: 'Something went wrong',
                 message: 'Please try again',
                 buttonText: 'Ok'
+            },
+
+            GEOLOCATION_REQUIRED: {
+                title: 'Something went wrong',
+                message: 'We’re sorry, but the restaurant does not currently deliver to this address. Please choose to order via collection or from another restaurant.',
+                buttonText: 'Back to order'
             },
 
             DuplicateOrder: {
@@ -186,7 +192,7 @@ const messages = {
     },
 
     serviceTypes: {
-        collection: 'Collection',
+        collection: 'Pick-up',
         delivery: 'Delivery'
     },