+++ packages/components/organisms/f-cookie-banner/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
 const LegacyCookieBanner = require('../../test-utils/component-objects/f-cookieBanner-legacy.component');
@@ -14,9 +13,7 @@ describe('Legacy Accessibility tests', () => {
         // Arrange
         const formattedLocale = 'en-AU';
         legacyCookieBanner.withQuery('&knob-Locale', formattedLocale);
-        const pageUrl = buildUrl(legacyCookieBanner.componentType, legacyCookieBanner.componentName, legacyCookieBanner.path);
-        legacyCookieBanner.open(pageUrl);
-        legacyCookieBanner.waitForComponent();
+        legacyCookieBanner.load();
 
         // Act
         const axeResults = getAccessibilityTestResults('f-cookie-banner');
@@ -29,9 +26,7 @@ describe('Legacy Accessibility tests', () => {
         // Arrange
         const formattedLocale = 'en-GB';
         cookieConsentBanner.withQuery('&knob-Locale', formattedLocale);
-        const pageUrl = buildUrl(cookieConsentBanner.componentType, cookieConsentBanner.componentName, cookieConsentBanner.path);
-        cookieConsentBanner.open(pageUrl);
-        cookieConsentBanner.waitForComponent();
+        cookieConsentBanner.load();
 
         // Act
         const axeResults = getAccessibilityTestResults('f-cookie-banner');