+++ packages/components/organisms/f-offers/test-utils/component-objects/f-offers.component.js
@@ -8,6 +8,10 @@ module.exports = class Offers extends Page {
 
     get component () { return $(COMPONENT); }
 
+    load () {
+        super.load(this.component);
+    }
+
     open (url) {
         super.open(url);
     }