+++ packages/components/organisms/f-takeawaypay-activation/test/visual/f-takeawaypay-activation-unauthenticated.visual.mobile.spec.js
@@ -1,0 +1,18 @@
+const TakeawayPayComponent = require('../../test-utils/component-objects/f-takeawaypayActivation.component');
+
+let takeawayPayComponent;
+
+describe('f-takeawaypay-activation - Unauthenticated - Mobile Visual Tests', () => {
+    beforeEach(() => {
+        // Arrange
+        takeawayPayComponent = new TakeawayPayComponent();
+
+        // Act
+        takeawayPayComponent.load();
+    });
+
+    it('should display the component when the user is not logged in.', () => {
+        // Assert
+        browser.percyScreenshot('f-takeawaypay-activation - Unauthenticated', 'mobile');
+    });
+});