+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,69 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.1.0
+------------------------------
+*October 5, 2021*
+
+### Changed
+- Increased maxBundleSize to 105Kb.
+
+v2.0.0
+------------------------------
+*October 5, 2021*
+
+### Changed
+- Updated version of `f-button`, `f-mega-modal`, `f-alert`, `f-card` and `f-form-field` to include icing phase 2 changes.
+- T&C, Privacy Policy and Cookie Policy links to use f-link component.
+
+### Removed
+- `o-link--noDecoration` class from the switch account link to add underline.
+- `is-rounded` prop from `f-card` component call as the card is now rounded by default.
+
+
+v1.6.0
+------------------------------
+*September 30, 2021*
+
+### Added
+- Analytics service file to handle analytics events.
+- `f-analytics` plugin to handle GTM events.
+
+
+v1.5.0
+------------------------------
+*September 30, 2021*
+
+### Changed
+- Only call Geolocation endpoint when the service type is delivery
+
+
+v1.4.0
+------------------------------
+*September 29, 2021*
+
+### Changed
+- `GEOLOCATION_REQUIRED` error handling.
+
+
+v1.3.0
+------------------------------
+*September 27, 2021*
+
+### Added
+- New function to map update checkout request
+
+### Changed
+- Map update checkout data before calling `handleUpdateCheckout`
+
+
+v1.2.0
+------------------------------
+*September 21, 2021*
+
+### Changed
+- Correct copy for collection time for AUZ
+
 
 v1.1.0
 ------------------------------